Added functionality to allow the User to control the query method (GET or POST) for a solrj query
This change updates the solrService.query() method to take an optional paremeter - SolrRequest.METHOD. If the user does not specify this parameter, the solrService uses the SolrRequest.METHOD that is stored in the solrService.queryMethod attribute. The solrService.queryMethod defaults to SolrRequest.METHOD.POST to support large complex queries by default.